Back to where i was before refactoring, guest/hypervisor stuff still needs a little love
